refactor(networking): extract SyncPolicyBase ABC — step 1 of #109
Extracts the full sync policy pipeline (circuit breaking, retry loop,
backoff, rate limiting, proxy rotation, robots.txt, metadata assembly)
into a new SyncPolicyBase ABC in _sync_policy_base.py.  HttpClient and
CurlHttpClient are now thin subclasses (~60 lines of actual logic each)
that supply a session, implement three abstract methods for exception
handling, and define the four public HTTP methods.

Resolves finding Q: urlparse(url).netloc computed once at the top of
_request and passed as `host` to _get_circuit_breaker and
_enforce_rate_limit, eliminating the duplicate parse per attempt.

Resolves finding R: _handle_request_exception timeout param is now
float | tuple[float, float] (no spurious | None).

Aligns HttpClient rate-limit semantics to end-to-start: timestamp is
now written in the _request finally block (matching the three other
clients), not in _enforce_rate_limit.  _enforce_rate_limit(host) now
takes the pre-computed host string rather than a full URL.

async_client._parse_retry_after: use targeted inline type-ignores to
survive both environments (one with full email.utils stubs, one without)
without relying on a bare unrestricted # pyright: ignore.

Test mock patch targets updated from ladon.networking.{client,curl_client}
to ladon.networking._sync_policy_base for sleep/uniform/monotonic.
All 532 tests pass. Pyright strict-mode: 0 errors.
